layout/base/tests/bug1550869-1c.html

<!DOCTYPE HTML>
<html class="reftest-wait">
  <title>Bug 1516963: Test AccessibleCaret doesn't show when clicking on an undraggable image button.</title>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <style>
  #container {
    border: 1px solid blue;
    width: 100px;
    height: 100px;
    background-color: yellow;
  }
  input {
    border: 0;
    outline: 0;
  }
  </style>
  <script>
  document.addEventListener("selectionchange", () => {
    ok(window.getSelection().isCollapsed, "The selection should be collapsed!");
  });

  function click() {
    let img = document.getElementById("img");
    synthesizeMouseAtCenter(img, {});
    setTimeout(() => { document.documentElement.removeAttribute("class"); });
  }
  </script>
  <body onload="SimpleTest.waitForFocus(click);">
    <div id="container"><input draggable="false" width="60px" height="40px" type="image" src="chrome/blue-32x32.png" id="img"></div>
  </body>
</html>
